[REFACTOR/#220] DiaryRepository 인터페이스와 구현체를 구현합니다.
