--- app/src/main/java/com/sopt/clody/data/repository/DailyDiariesRepository.kt
@@ -1,7 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.DailyDiariesResponseDto
-
-interface DailyDiariesRepository {
-    suspend fun getDailyDiariesData(year: Int, month: Int,date:Int): Result<DailyDiariesResponseDto>
-}
