--- app/src/main/java/com/sopt/clody/data/repository/DailyDiaryListRepository.kt
@@ -1,8 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.DailyDiariesResponseDto
-import com.sopt.clody.data.remote.dto.response.MonthlyCalendarResponseDto
-
-interface DailyDiaryListRepository {
-    suspend fun deleteDailyDiary(year: Int, month: Int, day: Int): Result<DailyDiariesResponseDto>
-}
