--- app/src/main/java/com/sopt/clody/data/repository/DiaryListRepository.kt
@@ -1,10 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.MonthlyDiaryResponseDto
-
-interface DiaryListRepository {
-    suspend fun getMonthlyDiary(
-        year: Int,
-        month: Int
-    ): Result<MonthlyDiaryResponseDto>
-}
