--- app/src/main/java/com/sopt/clody/data/repository/DiaryTimeRepository.kt
@@ -1,7 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.DiaryTimeResponseDto
-
-interface DiaryTimeRepository {
-    suspend fun getDiaryTime(year: Int, month: Int, date: Int): Result<DiaryTimeResponseDto>
-}
