--- app/src/main/java/com/sopt/clody/data/repository/MonthlyCalendarRepository.kt
@@ -1,8 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.base.ApiResponse
-import com.sopt.clody.data.remote.dto.response.MonthlyCalendarResponseDto
-
-interface MonthlyCalendarRepository {
-    suspend fun getMonthlyCalendarData(year: Int, month: Int): Result<MonthlyCalendarResponseDto>
-}
