--- app/src/main/java/com/sopt/clody/data/repository/ReplyDiaryRepository.kt
@@ -1,7 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.ReplyDiaryResponseDto
-
-interface ReplyDiaryRepository {
-    suspend fun getReplyDiary(year: Int, month: Int, date: Int): Result<ReplyDiaryResponseDto>
-}
