--- app/src/main/java/com/sopt/clody/data/repository/WriteDiaryRepository.kt
@@ -1,7 +1,0 @@
-package com.sopt.clody.data.repository
-
-import com.sopt.clody.data.remote.dto.response.WriteDiaryResponseDto
-
-interface WriteDiaryRepository {
-    suspend fun writeDiary(date: String, content: List<String>): Result<WriteDiaryResponseDto>
-}
